requirements.txt
deepl
feedparser
google-genai
packaging
python-dotenv
omegaconf
slack-sdk
